refactor: move auth api wrappers from utils/ to composables/useAuth.js
